+++ docs/CHANGELOG.md
@@ -12,6 +12,7 @@ Known [issues](https://github.com/vengi-voxel/vengi/issues?q=is%3Aissue+is%3Aope
 
 General:
 
+   - Added bone index save/load support to the `vengi` format (version 8)
    - Refactored parts of the io module and fixed memory leaks
    - Added support for `ldr` and `mpd` lego files
    - Improved `fbx` animation support and added write support